requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dotenv==0.19.0
requests==2.26.0
google-generativeai==0.3.0